+++ src/models/mongo/index.ts
@@ -1,7 +1,5 @@
 export * from "./AttachmentMongo.js";
 export * from "./AuditLogEntryMongo.js";
-export * from "./BookingMongo.js";
-export * from "./BookingTypeMongo.js";
 export * from "./BrandingMongo.js";
 export * from "./CalendarEventMongo.js";
 export * from "./CalendarShareLinkMongo.js";