+++ src/models/sql/index.ts
@@ -1,7 +1,5 @@
 export * from "./AttachmentSQL.js";
 export * from "./AuditLogEntrySQL.js";
-export * from "./BookingSQL.js";
-export * from "./BookingTypeSQL.js";
 export * from "./BrandingSQL.js";
 export * from "./CalendarEventSQL.js";
 export * from "./CalendarShareLinkSQL.js";