+++ src/routes/index.ts
@@ -4,8 +4,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 export * from "./BaseAttachmentRoute.js";
 export * from "./BaseAuditLogRoute.js";
-export * from "./BaseBookingRoute.js";
-export * from "./BaseBookingTypeRoute.js";
 export * from "./BaseBrandingRoute.js";
 export * from "./BaseCalendarEventRoute.js";
 export * from "./BaseCalendarShareLinkRoute.js";