+++ src/routes/mongo/index.ts
@@ -1,7 +1,5 @@
 export * from "./AttachmentRouteMongo.js";
 export * from "./AuditLogRouteMongo.js";
-export * from "./BookingRouteMongo.js";
-export * from "./BookingTypeRouteMongo.js";
 export * from "./BrandingRouteMongo.js";
 export * from "./CalendarEventRouteMongo.js";
 export * from "./CalendarShareLinkRouteMongo.js";